used safeERC20, and a change in Decimal Diff

+++ contracts/interfaces/IBimaPSM.sol
@@ -12,11 +12,7 @@ interface IBimaPSM {
     /// @param usbdAmount Amount of USBD tokens transferred
     /// @param timestamp Timestamp of calling the mint function
     event Mint(
-        address indexed from,
-        address indexed to,
-        uint256 underlyingAmount,
-        uint256 usbdAmount,
-        uint256 timestamp
+        address indexed from, address indexed to, uint256 underlyingAmount, uint256 usbdAmount, uint256 timestamp
     );
 
     /// @notice Emitted when a user redeems USBD tokens for underlying tokens
@@ -26,11 +22,7 @@ interface IBimaPSM {
     /// @param usbdAmount  Amount of USBD deposited for redeeming
     /// @param timestamp Timestamp of redeeming
     event Redeem(
-        address indexed from,
-        address indexed to,
-        uint256 underlyingAmount,
-        uint256 usbdAmount,
-        uint256 timestamp
+        address indexed from, address indexed to, uint256 underlyingAmount, uint256 usbdAmount, uint256 timestamp
     );
 
     /// @notice Emitted when there is not enough liquidity in the PSM
@@ -50,7 +42,7 @@ interface IBimaPSM {
     /// @notice Decimal numbers difference between the usbd and the underlying
     /// @dev Used for converting amounts between usbd and the underlying
     /// @return decimalDiff
-    function DECIMAL_DIFF() external returns (uint8 decimalDiff);
+    function DECIMAL_DIFF() external returns (uint256 decimalDiff);
 
     /// @notice Deposit underlying token and receive exact amount of USBD tokens in return 1:1
     /// @param _to Address that will receive USBD tokens
